+++ foundry/training/mlm.py
@@ -36,6 +36,7 @@ from typing import Callable, Optional
 import numpy as np
 
 from foundry.training._logger import _FoundryLogger
+from foundry.training._params import trainable_parameters as _trainable
 from foundry.training._scheduler import build_scheduler
 
 
@@ -188,7 +189,7 @@ class MLMTrainer:
     def _build_optimizer(self):
         import torch
         return torch.optim.AdamW(
-            self.student.parameters(),
+            _trainable(self.student),
             lr=self.cfg.learning_rate,
             weight_decay=self.cfg.weight_decay,
         )